enhancement 2, 3, and 4 complete
enhancment 2: changes the code so that the first entry is 1.
added start=1 to list and done function. created the variable
list_index = args.line_number - 1 in the do function. replaced
args.line_number with list_index in relevent positions. added an
if-statement which prints a friendly message to the user if they try
and do an item number that doesn't exist.

enhancement 3: prevent different indentation. in list and done numbers
are padded with leading zeros. This is limited to a million entries. I
will return later to improve this functionality

enhancement 4: implement deletion. i replicated the variable and
if-statement used in the do function so that users delete items
starting from 1 not 0 and receive a friendly print message if they try
to delete an item that doesn't exist or if an item is done.
i added delete to the choices.