enhancement 6 and 7 complete
enhancement 6: lists by priority. items are divided by those that are
ranked and those that are unranked, with those ranked appearing first.
relies on the existance of enhancement 7.

enhancement 7: assigning priority to an item. I added `pri` to choices
and changed the number of argumenents to accept multiple. used the
list_index variable to make sure the priority list starts at 1. throws
a friendly print message if two arguments are not passed by the user.
throws a friendly print message if the first argument is not a number
and the second is not a letter. throws a friendly print message if the
user assigns a priority to an item that does not exist.